+++ app/src/main/java/com/example/maplecal/ItemSelectedListener.kt
@@ -1,7 +1,5 @@
 package com.example.maplecal
 
-import com.example.maplecal.data.HyperData
-
 interface ItemSelectedListener {
     fun onHyperSelected(data: MutableList<HyperData>)
 }
