+++ app/src/main/java/com/example/maplecal/LoadingDialog.kt
@@ -1,4 +1,4 @@
-package com.example.maplecal.dialog
+package com.example.maplecal
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,7 +6,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.DialogFragment
 import com.bumptech.glide.Glide
-import com.example.maplecal.R
 import com.example.maplecal.databinding.DialogLoadingBinding
 
 class LoadingDialog : DialogFragment() {
