+++ app/src/main/java/com/example/maplecal/data/ArcaneLocalDataSource.kt
@@ -1,0 +1,4 @@
+package com.example.maplecal.data
+
+class ArcaneLocalDataSource {
+}
