+++ app/src/main/java/com/example/maplecal/data/HyperRepository.kt
@@ -1,0 +1,4 @@
+package com.example.maplecal.data
+
+class HyperRepository {
+}
