+++ app/src/main/java/com/example/maplecal/data/ParkRepository.kt
@@ -1,0 +1,4 @@
+package com.example.maplecal.data
+
+class ParkRepository {
+}
