+++ app/src/main/java/com/example/maplecal/data/SymbolRepository.kt
@@ -1,0 +1,4 @@
+package com.example.maplecal.data
+
+class SymbolRepository {
+}
