+++ app/src/main/java/com/example/maplecal/growth/GrowthViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.maplecal.growth
+
+class GrowthViewModel {
+}
