+++ app/src/main/java/com/example/maplecal/hyper/HyperViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.maplecal.hyper
+
+class HyperViewModel {
+}
