+++ app/src/main/java/com/example/maplecal/park/ParkResultRecyclerViewAdapter.kt
@@ -1,10 +1,10 @@
-package com.example.maplecal.adapter
+package com.example.maplecal.park
 
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.maplecal.data.ParkData
+import com.example.maplecal.ParkData
 import com.example.maplecal.databinding.ItemRecyclerParkResultBinding
 import java.text.DateFormat
 import java.text.DecimalFormat
