+++ app/src/main/java/com/example/maplecal/park/ParkViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.maplecal.park
+
+class ParkViewModel {
+}
