+++ app/src/main/java/com/example/maplecal/symbol/SymbolDialog.kt
@@ -1,4 +1,4 @@
-package com.example.maplecal.dialog
+package com.example.maplecal.symbol
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,30 +6,24 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.DialogFragment
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.maplecal.adapter.ParkResultRecyclerViewAdapter
-import com.example.maplecal.data.ParkData
-import com.example.maplecal.databinding.ParkDialogBinding
+import com.example.maplecal.SymbolData
+import com.example.maplecal.databinding.SymbolDialogBinding
 import com.example.maplecal.util.resize
 
-class ParkDialog : DialogFragment() {
-    private lateinit var binding: ParkDialogBinding
-    private lateinit var adapter: ParkResultRecyclerViewAdapter
+class SymbolDialog : DialogFragment() {
+    private lateinit var binding: SymbolDialogBinding
+    private lateinit var adapter: SymbolResultRecyclerViewAdapter
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = ParkDialogBinding.inflate(layoutInflater)
+        binding = SymbolDialogBinding.inflate(layoutInflater)
         arguments?.let {
-            it.getParcelableArrayList<ParkData>("park").toString()
+            it.getParcelableArrayList<SymbolData>("symbol").toString()
         }
         return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        initView()
     }
 
     override fun onStart() {
@@ -37,16 +31,21 @@ class ParkDialog : DialogFragment() {
         resize(0.9f, 0.9f)
     }
 
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        initView()
+    }
+
     private fun initView() {
-        initParkRecyclerView()
+        initSymbolRecyclerView()
         initButton()
     }
 
-    private fun initParkRecyclerView() {
-        adapter = ParkResultRecyclerViewAdapter()
+    private fun initSymbolRecyclerView() {
+        adapter = SymbolResultRecyclerViewAdapter()
         arguments?.let {
-            val data = mutableListOf<ParkData>()
-            val saveData = it.getParcelableArrayList<ParkData>("park")
+            val data = mutableListOf<SymbolData>()
+            val saveData = it.getParcelableArrayList<SymbolData>("symbol")
             if (saveData != null) {
                 for (i in 0 until saveData.size) {
                     data.add(saveData[i])
