+++ app/src/main/java/com/example/maplecal/symbol/SymbolRecyclerViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.maplecal.symbol
+
+class SymbolRecyclerViewModel {
+}
